+++ js/Extensions/CurrentDateIndication.js
