+++ js/Extensions/PriceIndication.js
@@ -8,8 +8,8 @@
  *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
  */
 'use strict';
-import H from '../../Core/Globals.js';
-import U from '../../Core/Utilities.js';
+import H from '../Core/Globals.js';
+import U from '../Core/Utilities.js';
 var addEvent = U.addEvent, isArray = U.isArray, merge = U.merge;
 /**
  * The line marks the last price from visible range of points.
