+++ js/Extensions/StaticScale.js
