+++ tools/gulptasks/jsdoc-classes.js
@@ -36,9 +36,9 @@ const SOURCE_GLOBS = [
     'js/Core/Axis/Tick.js',
     'js/Core/Time.js',
     'js/Core/Tooltip.js',
-    'js/parts-gantt/GanttChart.js',
-    'js/parts-gantt/TreeGridAxis.js',
-    'js/parts-gantt/TreeGridTick.js',
+    'js/Core/Chart/GanttChart.js',
+    'js/Core/Axis/TreeGridAxis.js',
+    'js/Core/Axis/TreeGridTick.js',
     'js/Core/Axis/ColorAxis.js',
     'js/Extensions/GeoJSON.js',
     'js/Maps/Map.js',
