+++ ts/Extensions/StaticScale.ts
